pyproject.toml
[tool.poetry]
name = "gs-careers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gs_careers"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
psycopg2 = "^2.9.5"
black = "^22.10.0"
isort = "^5.10.1"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
chatterbot = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
